- The Mapler client now has a login screen  - Added AES encryption to the client - server connection  - EXP points and such  - and more.. idfk D:
git-svn-id: https://subversion.assembla.com/svn/mapler-me@984 7cbc8f23-9b70-4854-b091-be65f742de0c

+++ branches/SomeFunnyDLL/Extensions.cs
@@ -41,7 +41,7 @@ namespace System
         }
         public static int CalculateChecksum(this byte[] pThis)
         {
-            return pThis.Sum(b => b);
+            return (int)pThis.Sum(b => (long)b);
         }
 
         public static bool HasFlag(this int pValue, int pFlag)
